+++ src/components/Image/index.ts
@@ -1,1 +1,1 @@
-export { default as Image } from "./Image";
+export { default as ImageHandler } from "./Image";