+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from "./components";
+export { addLazyLoading, removeLazyLoading } from "./utils/lazy-loading";