+++ src/utils/src-set.ts
@@ -1,6 +1,0 @@
-// <picture>
-//   <source srcset="dark-mode.png" media="(prefers-color-scheme: dark)" />
-//   <img src="light-image.png" />
-// </picture>
-
-// Blured image as backgrund image von <img /> darstellen